refactor: fix lint errors and update AuthController tests
