+++ test/auth.e2e-spec.ts
@@ -1,7 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { INestApplication } from '@nestjs/common';
 import request from 'supertest';
-import { AppModule } from './../src/app.module';
+import { AppModule } from '../src/app.module';
 import { getModelToken } from '@nestjs/mongoose';
 import {
   FastifyAdapter,
@@ -13,9 +12,9 @@ import {
   MockModel,
   mockModelFactory,
   mockQueryFactory,
-} from './../src/test-utils/mongoose-mock-factory';
+} from '../src/test-utils/mongoose-mock-factory';
 import * as argon from 'argon2';
-import { AuthResponseMessage } from './../src/auth/auth.enums';
+import { AuthResponseMessage } from '../src/auth/auth.enums';
 
 jest.mock('argon2');
